Move `onPlaceBlock` and `onBreakBlock`
